Rename trip/properties.js to trip/query.js
